feat: rename LenderSeekResource to LenderSeekAdmResource and update package structure
