+++ beribturing-facade/src/main/java/ing/beribtur/facade/api/feature/adm/lender/rest/LenderSeekAdmResource.java
@@ -1,9 +1,9 @@
-package ing.beribtur.facade.feature.admin.rest;
+package ing.beribtur.facade.api.feature.adm.lender.rest;
 
 import ing.beribtur.accent.message.QueryResponse;
 import ing.beribtur.accent.util.QueryResponseUtil;
 import ing.beribtur.aggregate.user.entity.Lender;
-import ing.beribtur.facade.feature.admin.command.FindDisabledLendersAdminQuery;
+import ing.beribtur.facade.api.feature.adm.lender.query.FindDisabledLendersAdmQuery;
 import ing.beribtur.feature.lender.seek.LenderSeek;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -15,15 +15,15 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/lender")
+@RequestMapping("/api/admin/lender")
 @RequiredArgsConstructor
-public class LenderSeekResource implements LenderSeekFacade {
+public class LenderSeekAdmResource implements LenderAdmSeekFacade {
     private final LenderSeek lenderSeek;
 
     @Override
-    @PostMapping("/find-disabled-lenders-admin/query")
+    @PostMapping("/find-disabled-lenders/query")
     @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public QueryResponse<List<Lender>> findDisabledLendersAdmin(FindDisabledLendersAdminQuery query) {
+    public QueryResponse<List<Lender>> findDisabledLendersAdmin(FindDisabledLendersAdmQuery query) {
         query.validate();
         int page = query.getPage();
         int size = query.getSize();
