--- beribturing-facade/src/main/java/ing/beribtur/facade/feature/admin/rest/LenderSeekFacade.java
@@ -1,11 +1,0 @@
-package ing.beribtur.facade.feature.admin.rest;
-
-import ing.beribtur.accent.message.QueryResponse;
-import ing.beribtur.aggregate.user.entity.Lender;
-import ing.beribtur.facade.feature.admin.command.FindDisabledLendersAdminQuery;
-
-import java.util.List;
-
-public interface LenderSeekFacade {
-    QueryResponse<List<Lender>> findDisabledLendersAdmin(FindDisabledLendersAdminQuery command);
-}
